requirements.txt
# 项目核心依赖
httpx==0.28.1
Pillow==9.5.0
playwright==1.45.0
tenacity==8.2.2
typer>=0.12.3
opencv-python
pydantic==2.5.2
python-dotenv==1.0.1
requests==2.32.3
parsel==1.9.1
pandas==2.2.3
cryptography>=45.0.7
psutil>=5.9.0
# 项目组件依赖
fastapi==0.110.2
uvicorn==0.29.0
flask>=2.3.0
# 数据处理依赖
jieba==0.42.1
wordcloud==1.9.3
matplotlib==3.9.0
# 可选依赖 - 仅在需要时安装
# aiomysql==0.2.0
# redis~=4.6.0
# aiofiles~=23.2.1
# sqlalchemy>=2.0.43
# motor>=3.3.0
# xhshow>=0.1.3
# openpyxl>=3.1.2
# pytest>=7.4.0
# pytest-asyncio>=0.21.0